// frontend/src/app/components/pages-sidebar/misure-cem/misure-cem-dto.ts
export class MisureCemDto {
    idimiscem?: any;
    sopralluogo?: string;
    tipologia?: string;
    regione?: string;
    comune?: string;
    provincia?: string;
    codsito?: string;
    data?: any;
    modstrum?: string;
    sonda?: string;
    tartstrum?: string;
    puntodimisura?: string;
    descpuntomis?: string;
    latitudine?: string;
    longitudine?: string;
    valeffrmsvm?: string;
    valeffrmsut?: string;
    permcontgio?: string;
    limleggevm?: string;
    limleggeut?: string;
    rangefreq?: string;
    gammamisura?: string;
    sovraccarico?: string;
    risoluzione?: string;
    misisotrop?: string;
}